+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Event emitting (pan, plot tile, image draw)
 ### Fixed
 - Tile image path
+- Mouse movement
 
 ## [2.2.0] - 2022-02-11
 ### Added